refactor(workflow-packs): isolate queue attention summary
